% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Depression Data set}
\format{
A data frame with 612 rows and 4 columns
\describe{
  \item{Depression}{A continous variable to value the level of depression}
  \item{Age}{A continous variable to value the age of participants}
  \item{Sex}{A demographic variable, 1 for female, 0 for male}
  \item{Fatalism}{A continous variable to value the fatalism of participants}
}
}
\usage{
data
}
\description{
Depression Data set
}
\keyword{datasets}
