% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter_depen_var.R
\name{deter_depen_var}
\alias{deter_depen_var}
\title{Determine the dependent variable and response variable}
\usage{
deter_depen_var(variable_vector, data)
}
\arguments{
\item{variable_vector}{a vector of variable names, cut from the original argument}

\item{data}{original dataset form users}
}
\value{
TRUE or Error Message
}
\description{
Check if the response variable and dependent variables could be find in
}
\examples{
variable_vector = c("Y","X1","X2")
colnames( dataset )= "Y"  "X1" "X2" "X3"
square(variable_vector,dataset)

}
