% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_linear_model.R
\name{generate_linear_model}
\alias{generate_linear_model}
\title{Generate the Linear Model}
\arguments{
\item{s}{A argument of your model, which is a formula of your function about response variable and dependent variable}

\item{data}{original dataset form users}
}
\value{
A coefficient table of this fitted linear model
}
\description{
The Main function of this package, it would calculate the fitted (Least Square Estimate) Linear Model from your argument and dataset
}
\examples{
generate_linear_model(Y~X1+X2+X3,data)

}
