% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta.R
\name{get_beta}
\alias{get_beta}
\title{Calculate Beta Matrix}
\usage{
get_beta(data)
}
\arguments{
\item{data}{the dataset that is already selected by the other function}
}
\value{
Beta Matrix, which is a matrix contains beta_0,beta_1, and so on.
}
\description{
From your dependent variable and response variable to calculate the beta matrix, or warning you that you input a dataset which is not full ranked
}
\examples{
get_beta(data)

}
