% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficient_table.R
\name{get_coefficient_table}
\alias{get_coefficient_table}
\title{Get Coefficient Table of your model}
\usage{
get_coefficient_table(beta_matrix, s)
}
\arguments{
\item{beta_matrix}{a coefficient matrix corresponding to dependent variable and intercept}

\item{s}{A formula of your linear model containing the dependent variable and response variable}
}
\value{
A data.frame, column names is Coefficient and Variable, each row corresponding to each Variables name and value of coefficient
}
\description{
Last step of a linear regression model, generate the coefficient and dependent vairable corresponding table
}
\examples{
get_coefficient_table(beta_matrix,Y~X1+X2+X3)

}
