% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depen_var.R
\name{get_depen_var}
\alias{get_depen_var}
\title{Get the variable}
\usage{
get_depen_var(s)
}
\arguments{
\item{s}{A formula of a linear regression model, contain the response variable and dependent variable, connect with +}
}
\value{
A vector of your variable name, start with response variable and dependnent variables following.
}
\description{
Get the dependent variable name and response variable name from the formula from the users input
}
\examples{
get_depen_var(Y~X1+X2+X3)

}
