% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPred.R
\name{get_pred}
\alias{get_pred}
\title{Get prediction of your model}
\usage{
get_pred(model, new_X)
}
\arguments{
\item{new_X}{A dataframe of the test set or unkonwn data point you want to predict by the fitted model}

\item{coefficient_table}{A dataframe with coefficient and dependent variables, generated by generate_linear_model function.}
}
\value{
A vector of response variable value corresponding to each new data point.
}
\description{
From the fitted model, using this function to calculate with the test set or unknown data point, valid for vector calculate.
}
\examples{
new_X <- data.frame(X1 = c(1, 2, 3), X2 = c(4, 5, 6), X3 = c(7, 8, 9))
model <- data.frame(Coefficients = c(-1,1,2,3),Variable = c("Intercept", "X1", "X2", "X3"))
get_pred(model,new_X)

}
