% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_test.R
\name{na_test}
\alias{na_test}
\title{Check N.A in your dataset}
\usage{
na_test(variable_vector, data)
}
\arguments{
\item{variable_vector}{A vector of variable from your formula argument, generate by get_depen_var function}

\item{data}{The dataset from you input}
}
\value{
A complete dataset with the variables selected, or an error that the dataset contains N.A
}
\description{
From the variable of your select from the input of argument, check if there is N.A in the data set of variable interested, and also return the data set of variable intereseted
}
\examples{
variable_vector <- get_depend_var(s)
data <- na_test(variable_vector,data)

}
