% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLinearModel.R
\name{generate_linear_model}
\alias{generate_linear_model}
\title{Generate the Linear Model}
\usage{
generate_linear_model(s, data)
}
\arguments{
\item{s}{A argument of your model, which is a formula of your function about response variable and dependent variable}

\item{data}{original dataset form users}
}
\value{
A coefficient table of this fitted linear model
}
\description{
The Main function of this package, it would calculate the fitted (Least Square Estimate) Linear Model from your argument and dataset
}
\examples{
df <- data.frame(Y = c(1, 2, 3, 4),X1 = c(2, 7, 4, 5),X2 = c(3, 4, 1, 6),X3 = c(4, 5, 6, 9))
generate_linear_model(Y~X1+X2+X3,df)

}
