% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula.R
\name{get_formula}
\alias{get_formula}
\title{Get the formula of your model}
\usage{
get_formula(model)
}
\arguments{
\item{coefficient_table, }{A dataframe, with coefficient and dependent variables.}
}
\value{
A complete mathematics formula of the fitted linear model
}
\description{
From the main funciton generate coefficient table, to generate the mathematics formula of this model
}
\examples{
model = generate_linear_model(Y~X1+X2+X3,data)
get_formula(model)

}
