% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pred.R
\name{get_pred}
\alias{get_pred}
\title{Get prediction of your model}
\usage{
get_pred(model, new_X)
}
\arguments{
\item{model}{A dataframe with coefficient and dependent variables, generated by generate_linear_model function.}

\item{new_X}{A dataframe of the test set or unkonwn data point you want to predict by the fitted model}
}
\value{
A vector of response variable value corresponding to each new data point.
}
\description{
From the fitted model, using this function to calculate with the test set or unknown data point, valid for vector calculate.
}
\examples{
dataset = LinearModelGenerator::data
model <- generate_linear_model(Depression~Age+Sex+Fatalism,dataset)
new_X <- data.frame(Age = c(15, 29, 34), Sex = c(1, 0, 1), Fatalism = c(2, 0, 5))
get_pred(model,new_X)

}
